requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.3
gunicorn==20.1.0
Pillow==9.2.0
requests==2.28.1
